# evcar-news/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/audit-logs": {
      "get": {
        "operationId": "AdminAudit_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 1000000,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Free text: matches action, entity type/id, actor e-mail label or request id.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityType",
            "required": false,
            "in": "query",
            "schema": {
              "example": "users",
              "type": "string"
            }
          },
          {
            "name": "entityId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": false,
            "in": "query",
            "description": "Exact action, or a prefix ending with \"*\" (e.g. \"auth.*\").",
            "schema": {
              "example": "auth.login_failed",
              "type": "string"
            }
          },
          {
            "name": "actorId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Inclusive lower bound (ISO-8601).",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Exclusive upper bound (ISO-8601).",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "-createdAt",
              "type": "string",
              "enum": [
                "createdAt",
                "-createdAt"
              ]
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AuditLogDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "422": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Search the audit log (newest first by default)",
        "tags": [
          "admin: audit"
        ]
      }
    },
    "/api/v1/admin/audit-logs/{id}": {
      "get": {
        "operationId": "AdminAudit_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AuditLogDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "One audit log entry",
        "tags": [
          "admin: audit"
        ]
      }
    },
...
  },
  "info": {
    "title": "EV Car News API",
    "description": "REST API of EV Car News (evcar.news). Conventions:\n- Prefix /api/v1; admin routes under /api/v1/admin (permission-guarded), personal routes under /api/v1/me.\n- Language: ?lang=ar|en overrides Accept-Language (default ar). Market: ?market=EG overrides X-Market.\n- Single resource: {\"data\": {...}}; lists: {\"data\": [...], \"meta\": {page, pageSize, total, totalPages}}.\n- Errors: {\"error\": {\"code\", \"message\" (localized), \"details\"?, \"requestId\"}}; validation → 422 VALIDATION_FAILED.\n- Money: {\"amount\": \"decimal string\", \"currency\": \"EGP\"}; missing values are null (never 0).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorBodyDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "VALIDATION_FAILED"
          },
          "message": {
            "type": "string",
            "example": "بعض الحقول غير صالحة. راجع التفاصيل."
          },
          "details": {
            "type": "object",
            "description": "Code-specific details (e.g. field errors for VALIDATION_FAILED)."
          },
          "requestId": {
            "type": "string",
            "example": "0b8f6c1e-5a1c-4c0e-9d6b-2f3b9b4d1a77"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ]
      },
      "ErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "PageMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 57
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
...
      "AuditActorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "displayName"
        ]
      },
      "AuditLogDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "actorId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "actor": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AuditActorDto"
              }
            ]
          },
          "actorLabel": {
            "type": "string",
            "nullable": true
          },
          "action": {
            "type": "string",
            "example": "users.roles.update"
          },
          "entityType": {
            "type": "string",
            "example": "users"
          },
          "entityId": {
            "type": "string",
            "nullable": true
          },
          "before": {
            "type": "object",
            "nullable": true,
            "description": "State before (secrets redacted)."
          },
          "after": {
            "type": "object",
            "nullable": true,
            "description": "State after (secrets redacted)."
          },
          "diff": {
            "type": "object",
            "nullable": true,
            "description": "{ field: { from, to } }"
          },
          "ip": {
            "type": "string",
            "nullable": true
          },
          "userAgent": {
            "type": "string",
            "nullable": true
          },
          "requestId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "createdAt",
          "actorId",
          "actorLabel",
          "action",
          "entityType",
          "entityId",
          "ip",
          "userAgent",
          "requestId"
        ]
      },
...
    }
  }
}